# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Konfidence API
  version: v1
  description: >
    HTTP gateway between CLI/UI clients and the Konfidence CRDs.
servers:
  - url: /api/v1
    description: Stage resources (versioned API)
paths:
...
  /auth/callback:
    get:
      operationId: authCallback
      summary: OIDC callback
      description: >
        Receives the authorization code from the IDP after successful login.
        Exchanges the code for tokens and establishes a session cookie.
      tags: [auth]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the application after successful authentication.
          headers:
            Set-Cookie:
              description: Contains the session ID.
              schema:
                type: string
            Location:
              description: Redirect url
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /identity:
    get:
      operationId: getIdentity
      summary: Get the current user identity
      description: Returns identity claims for the authenticated session.
      tags: [auth]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /exchange:
    post:
      operationId: postExchangeCode
      summary: Send temp exchange code and PKCE verifier to get session
      description: Validates code and verifier and returns session if valid.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                verifier:
                  type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kden_session
...
  schemas:
...
    Identity:
      type: object
      required: [name, givenName, familyName, email, roles]
      properties:
        name:
          type: string
        givenName:
          type: string
        familyName:
          type: string
        middleName:
          type: string
        email:
          type: string
        roles:
          type: array
          minItems: 1
          items:
            type: string
    Session:
      type: object
      required: [id]
      properties:
        id:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    Unauthorized:
      description: Authentication required or session expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"